move cryptomaterial module as an util submodule
